Moved the installed packages functionality from Profile class to PackageManager class. Should not have any behavior change.
