+++ apio/commands/apio_boards.py
@@ -161,7 +161,9 @@ def _list_boards_docs_format(apio_ctx: ApioContext):
 
     # -- Get the version of the 'definitions' package use. At this point it's
     # -- expected to be installed.
-    def_version, _ = apio_ctx.profile.get_installed_package_info("definitions")
+    def_version, _ = apio_ctx.package_manager.get_installed_package_info(
+        "definitions"
+    )
 
     # -- Collect the boards info into a list of entires, one per board.
     entries: List[Entry] = _collect_board_entries(apio_ctx)
