+++ apio/commands/apio_examples.py
@@ -124,7 +124,9 @@ def list_examples_docs_format(apio_ctx: ApioContext):
 
     # -- Get the version of the 'definitions' package use. At this point it's
     # -- expected to be installed.
-    def_version, _ = apio_ctx.profile.get_installed_package_info("definitions")
+    def_version, _ = apio_ctx.package_manager.get_installed_package_info(
+        "definitions"
+    )
 
     # -- Get list of examples.
     entries: List[ExampleInfo] = Examples(apio_ctx).get_examples_infos()
